requirements.txt
fastapi
uvicorn
streamlit

langgraph
langchain
langchain-community

chromadb

sentence-transformers

ollama

pymupdf
pandas
pytesseract
pillow

sqlalchemy
psycopg2-binary

python-dotenv


python-jose
passlib[bcrypt]

pymysql
